% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_realtime.R
\name{bind_realtime}
\alias{bind_realtime}
\title{bind_realtime}
\usage{
bind_realtime(stations = stations)
}
\arguments{
\item{stations}{station to pull data}
}
\description{
Check all realtime hydrometric stations in the Federal-Provincial network that exceed a user supplied time gap in records (defaults to the 60 minutes).
}
