% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_datamart_status.R
\name{check_datamart_status}
\alias{check_datamart_status}
\title{Check for difference in datamart reported stations}
\usage{
check_datamart_status(prov_terr_state_loc = NULL)
}
\arguments{
\item{prov_terr_state_loc}{Province, state or territory.}
}
\description{
Check for differences in datamart reported stations
}
\examples{
\dontrun{
check_datamart_status(prov_terr_state_loc = "PE")
}

}
