% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_realtime_lag.R
\name{check_realtime_lag}
\alias{check_realtime_lag}
\title{Check real time lag of realtime stations}
\usage{
check_realtime_lag(STATION_NUMBER = "ALL", PROV_TERR_STATE_LOC = "BC",
  data_interval = "hourly")
}
\arguments{
\item{STATION_NUMBER}{Station numbers to look at. Leave blank if all provincial stations are desired.
Currently you can't mix stations from two difference jurisdictions. See examples.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. Defaults to "BC". Will not accept ALL.}

\item{data_interval}{Examine hourly or daily data? Defaults to hourly}
}
\value{
\itemize{
\item STATION NUMBER Water Survey of Canada station number
\item time_obs Date and time of most recent observation from that station
\item time_mod Date and time of data file upload to datamart
\item time_lag Time difference (in hours) between \code{time_mod} and \code{time_obs}. This value represents
the time delay for the network from data acquisition to data upload.
}

rl_lag %>%
 filter(Lag < 10) %>%
   mutate(median = median(Lag)) %>%
   ggplot(aes(x = Lag)) +
   geom_histogram(bins = 100) +
   scale_x_continuous(breaks = seq(1,10,0.5)) +
   labs(x = "Lag Value (hours)",
        y = "Number of Occurences",
        title = "Realtime Hydrometric Station Data Lag",
        subtitle = "Each station in BC was pinged for the most recent hydrometric observation. The difference between this value and the time when the ping took place is the lag value.",
        caption = "Generated by Sam Albers - BC Ministry of Environment")
}
\examples{
check_realtime_lag(STATION_NUMBER = "08NL071")

## This will throw an error
\dontrun{
check_realtime_lag(STATION_NUMBER = c("08NL071","05QB002"))
}

## To check all stations in Alberta:
check_realtime_lag(PROV_TERR_STATE_LOC = "BC")

}
