% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stn_gap.R
\name{check_stn_gap}
\alias{check_stn_gap}
\title{Check for interval threshold value}
\usage{
check_stn_gap(stations = "ALL", gap_thres = 60, num_gaps = 5)
}
\arguments{
\item{stations}{select one station of all. Currently accepts either one station or "ALL"}

\item{gap_thres}{Threshold (in minutes) whereby there is a gap in the stations realtime data. Defaults to 60 minutes. DOES NOT CONNECT TO num_gaps}

\item{num_gaps}{Number of gaps that exceed 20 minutes. Defaults to 5.}
}
\value{
A dataframe containing all the stations (and associated spatial information) that exceed the gap criteria of gap_thres.
}
\description{
Check all realtime hydrometric stations in the Federal-Provincial network that exceed a user supplied time gap in records (defaults to the 60 minutes).
}
\examples{
library(HYDATaddons)
check_stn_gap(stations = "07EC003")

check_stn_gap(stations = "07EC003", gap_thres = 300)

}
