% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_water_office_status.R
\name{check_water_office_status}
\alias{check_water_office_status}
\title{A function to determine station status}
\usage{
check_water_office_status(station_number, print_url = FALSE,
  copy_url = FALSE)
}
\arguments{
\item{station_number}{WSC station number}

\item{print_url}{logical: Should the Water Office url be printed?}

\item{copy_url}{logical: Should the Water Office url be copied to a clipboard (windows only)?}
}
\description{
This function scrapes data off the ECCC water office and outputs a summary of station(s) status
}
\examples{
\dontrun{
check_water_office_status("08MF005",print_url = TRUE)

## to get around 20 station limit for the water office
if (require("purrr")) {
stns <- tidyhydat::realtime_stations(prov_terr_state_loc = "BC")
stns_split <- split(stns$STATION_NUMBER, (seq(length(stns$STATION_NUMBER))) \%/\% 20)
map_dfr(stns_split, ~ check_water_office_status(.x))
}
}

}
