% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate diagnostic report}
\usage{
generate_report()
}
\description{
run this command to render the Net_diag report. The reports are then outputted to the report folder
}
