% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrolook-package.R
\docType{package}
\name{hydrolook-package}
\alias{hydrolook}
\alias{hydrolook-package}
\title{hydrolook: Hydrometric Network Diagnostic}
\description{
Provides functions and reports to diagnose issues with the Water Survey of Canada Hydrometric network.
}
\author{
\strong{Maintainer}: Sam Albers \email{sam.albers@gov.bc.ca}

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
