% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrolook.R
\docType{package}
\name{hydrolook}
\alias{hydrolook}
\alias{hydrolook-package}
\title{Hydrometric Network Diagnostic}
\description{
This package provides functions and reports to diagnose issues with the Water Survey of Canada Hydrometric network

\code{hydrolook} package
}
