% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_diag_report.R
\name{net_diag_report}
\alias{net_diag_report}
\title{Generate diagnostic report}
\usage{
net_diag_report(output_type = "pdf", PROV_TERR_STATE_LOC = "BC")
}
\arguments{
\item{output_type}{the type of file to be outputted. Currently html and pdf are supported. defaults to pdf}

\item{PROV_TERR_STATE_LOC}{Province to be surveyed. Defaults to BC.}
}
\description{
run this command to render the Net_diag report. The reports are then outputted to the report folder
}
\examples{
\dontrun{
net_diag_report(output_type = "pdf", PROV_TERR_STATE_LOC = "PE")
}
}
\seealso{
Other report_generators: \code{\link{station_report}}
}
