% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_correlation.R
\name{network_correlation}
\alias{network_correlation}
\title{A function to map individual station correlations}
\usage{
network_correlation(station_number = NULL, cor_threshold = 0.6,
  method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{station_number}{Water Survey of Canada station number.}

\item{cor_threshold}{Threshold for correlation. Default is 0.6.}

\item{method}{From `stats::cor()` a character string indicating which correlation coefficient (or covariance)
is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.`}
}
\description{
This function makes use of the igraph and ggraph packages to create a network object from historical
hydrometric data. Still under construction.
}
\examples{
\dontrun{
network_correlation("08NM146")
}
}
