% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_lag_report.R
\name{realtime_lag_report}
\alias{realtime_lag_report}
\title{Generate diagnostic report}
\usage{
realtime_lag_report(output_type = "pdf", province = "BC")
}
\arguments{
\item{output_type}{the type of file to be outputted. Currently html and pdf are supported. defaults to pdf}

\item{province}{Province to be surveyed. Defaults to BC.}

\item{report_name}{Report to generate. Options:
\itemize{
\item Net_diag (not currently implemented)
\item Realtime_lag
}}
}
\description{
run this command to render the Net_diag report. The reports are then outputted to the report folder
}
\examples{
\dontrun{
realtime_lag_report(output_type = "pdf", province = "PE")
}
}
