% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_realtime_lag.R
\name{report_realtime_lag}
\alias{report_realtime_lag}
\title{Generate diagnostic report}
\usage{
report_realtime_lag(output_type = "pdf", PROV_TERR_STATE_LOC = "BC")
}
\arguments{
\item{output_type}{the type of file to be outputted. Currently html and pdf are supported. defaults to pdf}

\item{PROV_TERR_STATE_LOC}{Province to be surveyed. Defaults to BC.}
}
\description{
run this command to render the realtime report. The reports are then outputted to the report folder
}
\examples{
\dontrun{
realtime_lag_report(output_type = "pdf", PROV_TERR_STATE_LOC = "PE")
}
}
