% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_realtime_lag.R
\name{check_realtime_lag}
\alias{check_realtime_lag}
\title{Check real time lag of realtime stations}
\usage{
check_realtime_lag(stations = "ALL")
}
\examples{
rl_alg <- check_realtime_lag(stations = "ALL")

rl_lag \%>\%
 filter(Lag < 10) \%>\%
   mutate(median = median(Lag)) \%>\%
   ggplot(aes(x = Lag)) +
   geom_histogram(bins = 100) +
   scale_x_continuous(breaks = seq(1,10,0.5)) +
   labs(x = "Lag Value (hours)",
        y = "Number of Occurences",
        title = "Realtime Hydrometric Station Data Lag",
        subtitle = "Each station in BC was pinged for the most recent hydrometric observation. The difference between this value and the time when the ping took place is the lag value.",
        caption = "Generated by Sam Albers - BC Ministry of Environment")

}
