% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_realtime_lag.R
\name{check_realtime_lag}
\alias{check_realtime_lag}
\title{Check real time lag of realtime stations}
\usage{
check_realtime_lag(station_number = NULL, prov_terr_state_loc = NULL,
  data_interval = "hourly", tracker = FALSE)
}
\arguments{
\item{station_number}{Water Survey of Canada station number. No default. Can also take the "ALL" argument.
Currently you can't mix stations from two difference jurisdictions. See examples.}

\item{prov_terr_state_loc}{Province, state or territory. Defaults to "BC". Will not accept ALL.}

\item{data_interval}{Examine hourly or daily data? Defaults to hourly}

\item{tracker}{Should a progress list of stations be printed while the analysis is executed? Defaults to FALSE}
}
\value{
\itemize{
\item STATION NUMBER Water Survey of Canada station number
\item time_obs Date and time of most recent observation from that station
\item time_mod Date and time of data file upload to datamart
\item time_lag Time difference (in hours) between \code{time_mod} and \code{time_obs}. This value represents
the time delay for the network from data acquisition to data upload.
}
}
\description{
Check lag of real time stations using most recent observations and modification date of most recent file.
}
\examples{
\donttest{
check_realtime_lag(station_number = c("08NL071","05QB002"))

## To check all stations in PEI:
check_realtime_lag(prov_terr_state_loc = "PE")
}


}
