% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_report.R
\name{station_report}
\alias{station_report}
\title{Station report generator}
\usage{
station_report(output_type = "pdf", STATION_NUMBER = NULL)
}
\arguments{
\item{output_type}{the type of file to be outputted. Currently html and pdf are supported. defaults to pdf}

\item{STATION_NUMBER}{Water Survey of Canada station number. No default.}
}
\value{
Will output a report to a report/station_report directory
}
\description{
Commands to generate reports
}
\examples{
\donttest{
station_report(output_type = "pdf", STATION_NUMBER = "08EB005")
station_report(output_type = "pdf", STATION_NUMBER = "08MF005")
station_report(output_type = "pdf", STATION_NUMBER = "07EA005")
}


}
\seealso{
Other report_generators: \code{\link{net_diag_report}}
}
