% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_report.R
\name{station_report}
\alias{station_report}
\title{Station report generator}
\usage{
station_report(STATION_NUMBER, PROV_TERR_STATE_LOC)
}
\arguments{
\item{STATION_NUMBER}{Water Survey of Canada station number. No default.}

\item{PROV_TERR_STATE_LOC}{Province, state or territory. See also for argument options.}
}
\description{
Commands to generate reports
}
\examples{
\donttest{
station_report(STATION_NUMBER = "08MF005", PROV_TERR_STATE_LOC = "BC")
}


}
